updated navbar for new desktop layout
